+++ src/space/core/console_commands/events.rs
@@ -1,6 +1,6 @@
 use crate::space::core::networking::resources::ReliableServerMessage;
 
 pub struct NetConsoleCommands {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
