+++ src/space/core/entity/events.rs
@@ -1,21 +1,21 @@
 use crate::space::core::networking::resources::ReliableServerMessage;
 
 pub struct NetLoadEntity {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetShowcase {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetUnloadEntity {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetSendEntityUpdates {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
